Aim and Fire
I created a cursor system that can aim, and fires when the player pressed a specified button. The cursor has a line Renderer that tracks it, which we can use later to help communicate the player's powers. Later on, we will figure out how to control this cursor using the face