Adding Lives Counter
Now that the Zombie moves to a designated spot near the player, I'm trying to create a way for the zombie to do damage. To do this, I'm adding ints to track the player's health, and adding methods inside the zombie to subtract that health when triggered by the animation.