Adding in Animations
I've begun to use the StateTransitions to control the animation state of our enemy. I also now freeze the enemy when shot instead of deleting. Will have to find a way to change the material color, and see if we can make sure that no other enemies push into the frozen ones